add QingLong DTO classes for environment management
