+++ src/Ray.BiliBiliTool.Agent/QingLong/Dtos/AddQingLongEnv.cs
@@ -1,0 +1,8 @@
+namespace Ray.BiliBiliTool.Agent.QingLong.Dtos;
+
+public class AddQingLongEnv
+{
+    public string value { get; set; }
+    public string name { get; set; }
+    public string remarks { get; set; }
+}
