+++ src/Ray.BiliBiliTool.Agent/QingLong/Dtos/QingLongGenericResponse.cs
@@ -1,0 +1,8 @@
+namespace Ray.BiliBiliTool.Agent.QingLong.Dtos;
+
+public class QingLongGenericResponse<T>
+{
+    public int Code { get; set; }
+
+    public T Data { get; set; }
+}
