+++ src/Ray.BiliBiliTool.Agent/QingLong/Dtos/UpdateQingLongEnv.cs
@@ -1,0 +1,6 @@
+namespace Ray.BiliBiliTool.Agent.QingLong.Dtos;
+
+public class UpdateQingLongEnv : AddQingLongEnv
+{
+    public long id { get; set; }
+}
